requirements.txt
psycopg2-binary
appdirs==1.4.4
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.4.1
defusedxml==0.7.0rc2
distlib==0.3.1
Django==2.2.17
django-appconf==1.0.4
django-avatar==5.0.0
django-debug-toolbar==3.2
django-debug-toolbar-template-profiler==2.0.2
filelock==3.0.12
idna==3.1
oauthlib==3.1.0
Pillow==8.1.0
pip-tools==5.5.0
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
semantic-version==2.8.5
setuptools-rust==0.11.6
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
virtualenv==20.4.2
wrapt==1.12.1
